Hugh v2
Rather than requiring you to enter a connection string, Hugh v2 now
presents a QR code to pair with the controller.

This version also does end-to-end encryption, so the MQTT server doesn't
see the colour values.

It is vulnerable to replay attacks, it assumes the MQTT server is a
passive adversary (it's already trusted for availability) and not
actively spoofing messages.

The code is now split into three compartments.