Attempt to update config broker demo for sealed capabilities.
This is a pretty mechanical attempt at an update by analogy with
https://github.com/CHERIoT-Platform/network-stack/commit/95575b74b377d7b6b6d6aecab6b528c6c894662e
.  It is probably suboptimal: I suspect ConfigToken should really be
three types.  Also the macros in compartment-macros.h need updating to
better handle the situation where the declared type of the static
sealed thing differs from the actual type due to e.g. a trailing array
of variable length.