Config broker: fix two unintended crashes in parsers.
Both occurred when logging failure to parse a message so went unnoticed.
The logger parser was accidentally reading from a write-only capability due to using the wrong pointer.
The rgb_led parser was attempting to print a string that did not include the null terminator because the capability is resized to the result of strlen in order to emulate the real MQTT provider.
The latter is possibly something the debug printer should handle as per https://github.com/CHERIoT-Platform/cheriot-rtos/issues/450 .